CI: go back to macos-latest
Now that it seems we found a fix to #200, there is no reason to stick
to macos-11, which will likely be deprecated soon.

Update actions/setup-go to its latest version as well.
The new version uses caching by default, which we do not need.

While here, tidy up the cloneFile docs a bit.